+++ dashboard/views/main.py
@@ -23,12 +23,7 @@
 #    see http://www.gnu.org/licenses/.
 #
 ##############################################################################
-from django.contrib.auth.decorators import login_required, permission_required
-from .score_encoding import print_scores
-from base import models as mdl_base
-from django.contrib import messages
-from django.utils.translation import ugettext_lazy as _
-from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from base.views import layout
 
 
@@ -38,23 +33,3 @@ def home(request):
     return layout.render(request, "dashboard.html")
 
 
-@login_required
-@permission_required('base.is_tutor', raise_exception=True)
-def score_encoding(request):
-    return layout.render(request, "score_encoding.html", {})
-
-
-@login_required
-@permission_required('base.is_tutor', raise_exception=True)
-def download_papersheet(request):
-    person = mdl_base.person.find_by_user(request.user)
-    pdf = print_scores(request, person.global_id)
-    if pdf:
-        filename = "%s.pdf" % _('scores_sheet')
-        response = HttpResponse(content_type='application/pdf')
-        response['Content-Disposition'] = 'attachment; filename="%s"' % filename
-        response.write(pdf)
-        return response
-    else:
-        messages.add_message(request, messages.WARNING, _('no_score_to_encode'))
-        return score_encoding(request)
